refactor: 토너먼트 진행 페이지를 [id]/match로 이동
