+++ apps/web/src/app/tournament/[id]/match/_components/VsSection.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import type { ProductT } from '../_common/_types/tournament';
+import type { ProductT } from '../../_common/_types/tournament';
 import useCardSelectionAnimation from '../_hooks/useCardSelectionAnimation';
 import FinalProductCard from './FinalProductCard';
 import ProductCard from './ProductCard';
