+++ apps/web/src/app/tournament/[id]/match/page.tsx
@@ -5,10 +5,10 @@ import { notFound, redirect } from 'next/navigation';
 import { getQueryClient } from '@/utils/queryClient';
 
 import { postStartTournament } from './_apis/postStartTournament';
-import { getTournament } from './_common/_apis/getTournament';
-import type { GetTournamentInProgressResponseT } from './_common/_types/tournamentResponse';
-import { parseTournamentId } from './_common/_utils/parseTournamentId';
 import TournamentClient from './_components/TournamentClient';
+import { getTournament } from '../_common/_apis/getTournament';
+import type { GetTournamentInProgressResponseT } from '../_common/_types/tournamentResponse';
+import { parseTournamentId } from '../_common/_utils/parseTournamentId';
 
 type TournamentPageProps = {
   params: Promise<{ id: string }>;
